# agent-team-workbench/contracts/web/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Team Workbench Web API
  version: 0.1.0
  description: |
    浏览器 ↔ Go Control Plane 协议（协议文档 §5）。
    约定：/api/v1 前缀；资源 ID 带类型前缀（ws_/agent_/wi_/run_/approval_/artifact_）；
    JSON 字段 snake_case；写命令要求 Idempotency-Key；可变资源带 version/ETag；
    错误一律 application/problem+json。
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /workspaces/{workspace_id}:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    patch:
      tags: [workspace]
      operationId: updateWorkspace
      description: 更新名称、时区和策略；Owner/Admin。
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkspacePatch" } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Workspace" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /runs/{run_id}/commands/input:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: sendRunInput
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instruction]
              properties:
                instruction: { type: string }
      responses:
        "202": { description: 命令已接受 }
...
  /runs/{run_id}/commands/resume:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: resumeRun
      description: 仅在能力声明 resume 时可用，否则 422。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "202": { description: 命令已接受 }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
  /workspaces/{workspace_id}/runtime-bindings:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [runtime-bindings]
      operationId: listRuntimeBindings
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/RuntimeBindingList" } } } }
    post:
      tags: [runtime-bindings]
      operationId: createRuntimeBinding
      description: 添加 Runtime 与模型配置（provider/model/CredentialRef）；凭据只存引用。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [runtime_label, adapter_id]
              properties:
                runtime_label: { type: string }
                adapter_id: { type: string }
                provider: { type: string }
                model: { type: string }
                credential_ref: { type: string }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/RuntimeBinding" } } } }
...
  /runtime-bindings/{runtime_binding_id}/commands/probe:
    parameters: [{ $ref: "#/components/parameters/RuntimeBindingId" }]
    post:
      tags: [runtime-bindings]
      operationId: probeRuntimeBinding
      description: 探测版本、认证、协议与能力；不启动业务 Run。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "202": { content: { application/json: { schema: { $ref: "#/components/schemas/ProbeResult" } } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    WorkspaceId: { name: workspace_id, in: path, required: true, schema: { type: string } }
...
    RunId: { name: run_id, in: path, required: true, schema: { type: string } }
...
    RuntimeBindingId: { name: runtime_binding_id, in: path, required: true, schema: { type: string } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: |
        写命令幂等键。同 key 同请求体：已完成则重放原响应（Idempotent-Replayed: true），
        执行中返回 409 idempotency_in_progress（retryable，稍后重试）；
        同 key 不同请求体返回 409 idempotency_conflict；5xx 后占位释放，可同 key 重试。
...
  responses:
    Conflict:
      description: 幂等冲突或乐观锁版本冲突
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Unprocessable:
      description: 领域校验失败（状态机非法迁移等）
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
        retryable: { type: boolean }
        current_version: { type: integer }
...
    Workspace:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        timezone: { type: string }
        version: { type: integer }
    WorkspacePatch:
      type: object
      properties:
        name: { type: string }
        timezone: { type: string }
        expected_version: { type: integer }
...
    RuntimeBindingList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/RuntimeBinding" } }
    RuntimeBinding:
      type: object
      properties:
        id: { type: string }
        runtime_label: { type: string }
        adapter_id: { type: string }
        adapter_version: { type: string }
        provider_version: { type: string, nullable: true }
        provider: { type: string, description: "模型提供方，如 deepseek / openai" }
        model: { type: string, description: "模型配置，如 deepseek-chat" }
        credential_ref: { type: string, description: "CredentialRef；绝不返回凭据明文" }
        capabilities: { type: object, additionalProperties: { type: string } }
        status: { type: string, enum: [ready, probing, degraded, unavailable] }
        version: { type: integer }
    ProbeResult:
      type: object
      properties:
        runtime_binding_id: { type: string }
        ok: { type: boolean }
        provider_version: { type: string, nullable: true }
        capabilities: { type: object, additionalProperties: { type: string } }
        error: { type: string, nullable: true }
...